Create function pointer
In this code, I used function pointer for the first time.
I developed a calculator that calculates operations such as addition and division.... 
Try all the options. 
For fun, try to choose a number that is different from the existing options. 
Make this choice more than once and you will discover some surprises😂
Try to choose a number and not a letter, because I am too lazy to give an answer if the choice is a letter. I am sorry for my laziness.😂
If you have anything to add to make my code better and smaller, please give it to me because I think my code could be smaller than this.❤
My whatssap number : +212615928493
My instagram :
Mahdi_shaimi